agent_ai/requirements.txt
flask
pandas
python-dotenv
PyPDF2
openai